move testing binary utils into separate file
